Move MockIOFile to mocks util file
